RN-501: Moved OptionModel to @tupaia/database
